AICORE-612: Check for base versions when enriching & improve logs (#597)
We have encountered a NullPointerException within SaveEnrichmentFunction.
This was a result of corrupted documents where the documents did not have
base versions. This change makes an improvement to the class. We now
check if the base version is null and if it is null and the document is
not checked out, we exit, returning null in the function. Additionally,
this change improves logging within the class, adding an error log for
unexpected exceptions along with a couple other logs. The class also had
a couple small refactors as well.