AICORE-616: Resolve ComprehendService bug, skip text over quotas (#607)
We utilize Amazon Comprehend within our ComprehendService. There are
quota limits that are enforced when detecting entities, key-phrases,
dominant languages, sentiment, targeted sentiment, and syntax.
Previously, if we tried to analyze text with a text length beyond the
quota, we would encounter a TextSizeLimitExceededException. This change
skips Amazon Comprehend analysis when text is larger than the specified
quotas. In the future, we may implement another solution.

This is a forward port from master-10.10. For reference, see commit
6e84c5ca and pull request #604.