AICORE-626: Make NuxeoCloudClient dataset uploads more resilient (#633)
We have encountered a NullPointerException within the
NuxeoCloudClient, when creating and uploading dataset
batches. A blob file was null, causing the program to run
into into a NullPointerException within the FileBlob
constructor. This commit makes the NuxeoCloudClient more
resilient. We now check for null blob files within the
createBatch() and uploadDataset() methods.

This is a forward port from #627 and #629 for 2023.