NXBT-3895: Fix Dependabot access to maven-internal private registry
In fact, this repository is still needed to fetch old artifacts.
Thus, we need to configure it as a private registry in the Dependabot configuration. Otherwise, the Dependabot version update jobs are unable to authenticate to this registry, complaining with:

```
private_source_authentication_failure {source: "https://packages.nuxeo.com/repository/maven-internal"}
```